backend/pyproject.toml
[project]
name = "rgp-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    # Core
    "django>=5.2,<6.0",
    "wagtail>=7.0,<8.0",
    # API / headless
    "djangorestframework>=3.16,<4.0",
    "django-cors-headers>=4.7,<5.0",
    # Database
    "dj-database-url>=2.3,<3.0",
    # Static files
    "whitenoise>=6.9,<7.0",
    # Env config
    "python-decouple>=3.8,<4.0",
    # Images
    "Pillow>=11.0",
    # Production DB
    "psycopg2-binary>=2.9",
]

[tool.uv]
dev-dependencies = [
    "django-extensions>=3.2",
    "ipython>=8.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = []
